add feature to debug print in Merkle construction
We just print the size of each Merkle tree that is constructed and
when the construction is done.

This is like a 0-th order benchmark (*at least* 50% of the time is
spent constructing Merkle trees when running `prove_singles`, I think
more). But more importantly we now know the actual sizes for which we
construct Merkle trees!